move router evm server code to its own file
